+++ packages/js/src/Modules/Verto/util/interfaces.ts
@@ -25,6 +25,7 @@ export interface IVertoOptions {
   debugOutput?: 'socket' | 'file';
   prefetchIceCandidates?: boolean;
   forceRelayCandidate?: boolean;
+  trickleIce?: boolean;
   keepConnectionAliveOnSocketClose?: boolean;
   region?: string;
   /**
@@ -35,6 +36,12 @@ export interface IVertoOptions {
     target_id: string;
     target_version_id?: string;
   };
+
+  /**
+   * rtcIp & rtcPort options
+   */
+  rtcIp?: string;
+  rtcPort?: number;
 }
 export interface SubscribeParams {
   channels?: string[];