+++ packages/js/CHANGELOG.md
@@ -1,3 +1,29 @@
+## [2.24.1](https://github.com/team-telnyx/webrtc/compare/webrtc/v2.22.19...webrtc/v2.24.0) (2025-10-30)
+
+- fix: update inner `voice_sdk_id` state on first trickle ice canary use (#459)
+
+## [2.24.0](https://github.com/team-telnyx/webrtc/compare/webrtc/v2.22.19...webrtc/v2.24.0) (2025-10-29)
+
+### Features
+
+ - true trickle ice implementation (#439)
+
+### Enhancements
+
+ - check for gateway trickle ice support and add fallback
+ - improve logging on connection state changes
+ - add trickle ice performance metrics
+ - remove legacy video code
+
+## [2.22.19](https://github.com/team-telnyx/webrtc/compare/webrtc/v2.22.18...webrtc/v2.22.19) (2025-10-29)
+
+fix: enforce callID from call options to be string (#456)
+
+## [2.22.18](https://github.com/team-telnyx/webrtc/compare/webrtc/v2.22.17...webrtc/v2.22.18) (2025-10-23)
+
+fix: expose getIsRegistered (#448)
+fix: preferred codecs should only be enforced on INVITE, not on ANSWER (#452)
+
 ## [2.22.17](https://github.com/team-telnyx/webrtc/compare/webrtc/v2.22.16...webrtc/v2.22.17) (2025-08-04)
 
 chore: add detailed performance measuring for invite (#438)